+++ src/components/ProductHeaderRegistBtn.js
@@ -5,7 +5,7 @@ import "./ProductHeaderRegistBtn.css";
 
 // 이미지
 
-function ProductHeaderRegistBtn() {
+function ProductHeaderRegistBtn( ) {
   return <button className="registerBtn"><Link to ="/register">상품 등록하기</Link></button>
 
 }
