+++ src/components/common/RegistrationInput.css
