+++ src/components/common/RegistrationInput.js
