+++ src/pages/Marketpage.js
@@ -3,15 +3,15 @@ import { useState } from "react";
 import "./Marketpage.css";
 
 //렌더링 컴포넌트
-import ProductHeaderRegistBtn from "../components/ProductHeaderRegistBtn";
-import ProductHeaderSearchBar from "../components/ProductHeaderSearchBar";
-import ProductHeaderSortBtn from "../components/ProductHeaderSortBtn";
-import ProductHeaderText from "../components/ProductHeaderText";
+import ProductHeaderRegistBtn from "../components/ProductHeaderRegistBtn.js";
+import ProductHeaderSearchBar from "../components/common/ProductHeaderSearchBar.js";
+import ProductHeaderSortBtn from "../components/common/ProductHeaderSortBtn.js";
+import ProductHeaderText from "../components/common/ProductHeaderText.js";
 
 // 렌더링 프레임
-import HomepageRenderFooter from "../frame/HomepageRenderFooter.js";
-import HomepageRenderHeader from "../frame/PageHeaderRender.js";
-import ProductRenderGrid from "../frame/ProductRenderGrid.js";
+import HomepageRenderFooter from "../components/HomepageRenderFooter.js";
+import PageNavRender from "../components/PageNavRender.js";
+import ProductRenderGrid from "../components/ProductRenderGrid.js";
 
 // 커스텀 훅
 import useProductData from "../hooks/useProductData.js";
@@ -46,7 +46,7 @@ function Marketpage() {
   return (
     <div className="marketpage">
       <nav>
-        <HomepageRenderHeader marketBoardActive={true} loginStatus={false} device={Device}/>
+        <PageNavRender marketBoardActive={true} loginStatus={false} device={Device}/>
       </nav>
 
       <main>
